src/backend/Archaios.AI.GeeProcessor/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
geemap==0.20.6
earthengine-api>=0.1.332
matplotlib>=3.5.0
pillow>=9.0.0
setuptools>=65.0.0
azure-storage-blob>=12.13.0
numpy>=1.24.4
rasterio>=1.3.9